+++ build-tools/packages/build-tools/src/common/biomeConfigUtils.ts
@@ -10,7 +10,7 @@
 import { readFile } from "node:fs/promises";
 import * as JSON5 from "json5";
 
-import type { GitRepo } from "./gitRepo";
+import type { GitRepo } from "./gitRepo.js";
 
 // find-up is an ESM-only package and we're still building CJS, so this import form is required
 // so that TypeScript won't transpile it to require(). Once this package is switched to ESM, then